Data.Digest.Murmur32

Description

Type class and primitives for constructing 32 bit hashes using the MurmurHash2 algorithm. See http://murmurhash.googlepages.com for details on MurmurHash2.

Documentation

data Hash32

A 32 bit hash.

asWord32 :: Hash32 -> Word32

Extract 32 bit word from hash.

class Hashable32 a where

Methods

hash32Add :: a -> Hash32 -> Hash32

hash32AddWord32 :: Word32 -> Hash32 -> Hash32

hash32AddInt :: Int -> Hash32 -> Hash32

hash32 :: Hashable32 a => a -> Hash32

Create a hash using the default seed.

hash32WithSeed :: Hashable32 a => Word32 -> a -> Hash32

Create a hash using a custom seed.

The seed should be non-zero, but other than that can be an arbitrary number. Different seeds will give different hashes, and thus (most likely) different hash collisions.